Add password hashing functionality and update main script to use it
Implemented a new function `_hash_password` in the `auth.py` module to securely hash user passwords using the bcrypt library. This function takes a plaintext password as input, encodes it to bytes, generates a salt, and then hashes the password using the salt. The resulting hashed password is returned as bytes.

Updated the `main.py` script to import and demonstrate the use of the `_hash_password` function, replacing the previous example code that interacted with the database. This change showcases the new password hashing functionality, enhancing the security of user authentication processes in the application.