Add reset password token functionality to user authentication service
This commit introduces a new method `get_reset_password_token` in the `Auth` class
of the user authentication service. The method generates a reset token for a user
based on their email address. If the user is found, a UUID is generated and assigned
as the user's reset token, updating the corresponding field in the database. If the
user does not exist, a `ValueError` is raised. This functionality enhances the service
by providing a secure and straightforward way for users to reset their passwords.